Bugfix: The "-mode" functionality is incomplete and incorrect in the bootstrap.sh script.
The bootstrap_load_settings_has() method is removed for being incorrect.

Add missing Objects, such as build_indexer, build_indexer_arguments, build_name, search_exclusive, and search_static, to the "has-" detection data array.
Add additional detection data array for the "-mode" variants of the "has-" data array keys.

Stop checking if Content is an empty string when doing an Object existence check.
Just check to see if the Object exist, with or without Content.

Override the parameters when a "-mode" is provided for single-value parameters, even if the Content does not exist.
This should properly fail for when the Object is overridden with a mode that has no Content and is required.

For example, take the following:
  build_compiler gcc
  build_compiler-clang

In that above case clang is a mode that is overriding but has no Content.
The build_compiler is required.
This should fail because clang mode replaces the single-valued build_compiler with no Content.
This is a proper failure.