Update: De-couple the build settings data file from the install.sh script.
The design is originally focused around the data/build/settings file.

This file is simple enough and easy enough to use that I decided to use it beyond its intended design.

I noticed that the data/build/settings isn't required anymore.
Removing this allows me to better use in other projects, such as those that have only a fakefile or have a different directory location for the build settings data.

The install.sh script is even simpler now.