Feature: Support passing a fakefile Section name to the main program.
This is intuitive and should have already been supported in the project to better transition from GNU Make into Featureless Make.

The unknown parameter error is to no longer be thrown.

Instead, when the make process in being followed, check to see if any Sections exist in the fakefile matching the main program Section parameters.
Fail when this happens.
If the Section exists in the fakefile then that Section is operated on.

This, for example, now allows for doing something like "fake install" in the same way "make install" works.

The "install" Section must exist in the fakefile for this to work.