Bugfix: Support for passing a fakefile Section incorrectly fails.
The commit 25671e39d0a275b4a14b1db6612ae46d497c38c7 added a loop around the code.
I forget to handle the return status and break out of the loop appropriately.

Move the variable declaration outside of the loop to avoid re-declaring it inside.

An error or a child signal should cause the loop to terminate.

When the string comparison is matched, do not continue processing the inner loop.