Bugfix: Support for passing a fakefile Section incorrectly fails.
The commit 6c05c66f50719c9edaa36dea75669e03cd558ee4 added a loop around the code.
I forget to handle the return status and break out of the loop appropriately.

Move the variable declaration outside of the loop to avoid re-declaring it inside.

An error or a child signal should cause the loop to terminate.

When the string comparison is matched, do not continue processing the inner loop.