Feature: Add missing documentation handling functionality to Featureless Make.
This is an oversight and should have been implemented before the first stable release.
This feature should be a core feature.

Documentation is important.
Make sure the handling of documentation is available for use.

This adds "build_sources_documentation" to the build settings to help facilitate documentation installation.
The documentation that is copied is expected to be under 'data/documentation/'.

The 'documentation' and 'documents' terminology overlaps to some degree.
The 'documentation' is a more specific thing and may not always be a 'document'.
The existing behavior around 'documents' is left unchanged.

The skeleton operation now creates the 'data/documentation/' directory.

The bootstrap.sh script now supports enabling and disabling documentation via the more common '--enable-doc' and '--disable-doc' parameters.