Update: Improve logic and design of install.sh script.
The destination directories should be created if they do not already exist.
Fail when the destination exists but is not a directory rather than fail if that directory does not exist.

Fix grammar in 'Failed to create work directories'.

The includes should be created only when work is not an empty string.
The recently added documentation needs to be created as well.
The settings is not created and needs to be created just like with documentation.

Ensure the existence of these directories allows the copy and paste to work correctly.
Ensure the normal directories get created in addition to the work directories.

I originally designed this to not have any opinions for the settings.
I've decided to relax that and apply the default settings of the 'etc/' directory.
This directory is now installed to.
If say the 'etc/' is actually needing to be the 'share/' directory, custom paths can be altered as well.

Make sure the documentation and settings can be disabled or have their destination paths altered.