Cleanup: Fix license confusion.
The documentation should be under the cc-by-sa-4.0 license.
However, the specification must be under the open-standard-license-1.0 license.

This file is documentation on the time specification.
If the documentation is under another license, the standard it describes must still be followed according to the open-standard-license-1.0 license.
The documentation would have to communicate that the standard it is describing is still under the open-standard-license-1.0 license even if the documentation itself is under a different license.
This creates a confusing and misleading situation.
Clarify the situation by setting the license to open-standard-license-1.0.

Setting the license explicitly to open-standard-license-1.0 should hopefully address any misunderstandings.