Progress: Continue program related work, focusing primarily on fake and console parameter handling.
I realized that the new callback console parameter needs to be implemented.
I reviewed the drafted design and changed it accordingly.
I've been able to reduce the need to have a callback for each parameter and have a single callback for the set of parameters.

The documentation for the console parameter callback change is incomplete for this and my next progress should complete this documentation.

I updated several programs regarding the console changes and noticed a small number of bugs in the UTF8 program.
The first line is now being printed, conditionally, but future work may be needed.