Bugfix: FSS Write programs don't correctly handle multi-Content per Object situations.
The FSS Extended should support one or more Content per Object.
All others support the same number of Objects as Content (not referring to nested Content).

The error and error message when this is and is not correctly being handled.

The correct detection needs to count based on sub locations rather than locations.

Remove now pointless and misleading checks.