Bugfix: Fakefile processing with color context is producing extra spaces when unquoted.
Do not generate a new argument when calling fake_make_operate_expand_context().
The color context should combine with existing arguments.
This does not handle the last context argument and more work is likely needed to handle that one.

Remove the not quoted check that always generates a new argument.