Bugfix: Further Fakefile processing with iki replacement space problems.
This is a follow up to commit 3713a5f0bd90d32f564afaca29f01d9ccf329cfb.

The other IKI variables still need the new lines when the context does not.

Detect when certain IKI variables expand into nothing and in these cases do not append.
When performing last line, only append if it is not by itself.