Update: Add fanalyzer build mode.
I keep forgetting to run the GCC fanalyzer sanity checks to help catch problems before making a release.
Add a formal mode in all of the build settings to make this processor easier and therefore easier to remember.

Add missing -Wall to some of the files for the test mode.