Progress: Continue program related work, with emphasis on Featureless Make.
The common.h and related files are now broken up a little more.
There is now a common-string.h and common-type.h (with matching source files).
This should make it easier to navigate, particularly when there are a lot of functions, strings, and types.

Lots of tedious updates to the Featureless Make.
There are still a lot of structural changes to make.
Add more parameters to Featureless Make to address the documents and the licenses directories.
There is still more work to do with those parameters.

Fix some observed mistakes in the level_0 console code.