Bugfix: The build settings 'path_sources' is incorrectly added.
The commit 79858cf386916fd25ccd3578797de53b16f5c446 fixed a problem with the sources path passed to the program.
This revealed a bug where the build setting 'path_sources' is being added twice.
This is probably why the previous bug was not noticed for so long.

Make sure to apply the passed path sources parameter when the build settings does not have 'path_sources'.