Bugfix: The bootstrap script "-d/--define" command does not match the Fake command.
This is the old way the -d/--define command was designed.
The bootstrap script simply has never been updated.

The command is now "--define" and not "--defines".

Remove the sanity check in the parameter.
This behavior is not as precise and the commands are passed to the program in whatever form they expect.