Bugfix: Multiple single-valued Objects should used last instead of first Object.
Early on I was uncertain on what the proper behavior should be.

Over time I established that the last Object specified should be used.
This makes sense with most things.
Consider normal math:
  x = 1
  x = 2

Is x equal to 1 or 2?
I strongly believe the most common interpretation would be "2" because the assignment to "2" comes after the assignment to "1" when reading top-down.

Following this logic, the single-valued settings Objects must use the last occurrence rather than the first.

This is being considered a bug.