Progress: Continue working on Featureless Make, focusing on printing.
I noticed some odd behavior during this process where the file size of the resulting binary is increasing by as little as 72 bytes per function created.
I experimentally created a "void example_print(void) { }" function and was able to get that 72 number.
Fully populating or using the example_print function did not reduce or noticeably increase the file size.

I gave up experimenting and decided to continue on.
When I got close to wrapping up the print functions in the make, the resulting binary file size suddenly dropped 10k or so at least.
Once that dropped happened the functions no longer appeared to increase file size.
I suspect that there was some sort of weird optimization going on.

Some of the generic print functions in the fake make will probably be moved into the common directory as common print functions.