Bugfix: Incorrect and messy file error messages.
The code is unlikely to get to the error cases that print these messages due to other safety checks.
This makes testing this more difficult.
I do not currently have the testing environment for this.

I found the error messages is messed up and incorrect.
Clean up the messages.
This has not been tested due to the above mentioned problems.