Progress: Continue program work, wrapping up side-track work and diving back into fake.
Update projects to have a print directory in a similar manner as a common directory.

The previously completed and working programs should now be working again.
The side-track changes that I set out to complete in regards to them is now complete.

There is a huge amount of work ahead in regards to Featureless Make.
Continue forward step by step.