Bugfix: Add missing includes, add missing fanalyzer mode, and fix problems with stand alone fake.
The stand alone Featureless Make is not building properly.
Investigation reveals that there are missing sources and missing header includes.

For some reason the missing includes do not cause a problem in 0.6 or in non-stand_alone 0.7 builds.
However, it does cause a problem in the stand alone build.
Add the missing header includes.