Progress: Continue program related work, focusing on Featureless Make and printing.
Make the fl_print structure more extensible by adding a flag and a custom object.

The first/last printing is to complex.
Simplify the design to a simpler one.

Fix other problems as observed.

I have observed a memory leak in the recent f_directory_listing_t processing code.
Specifically, the f_directory_recurse_t as called in fake_build_copy().
The appropriate memory deletes appear to be being called.
More investigation is needed.
This is a recursive function, so there is probably a bad pointer assignment somewhere.