Update: Add fl_directory_do(), add new status codes, update fl_directory_copy(), and other changes.
An alternative to the ftw() is desired.
The function, fl_directory_do() is provided to perform generic recursive tasks.
This is based off of the fl_directory_copy() function.
This is not yet tested and there may be changes in the future once this is tested.

I noticed some things that could be improved in fl_directory_copy().
The recursion structure is actually specific to fl_directory_copy(), so rename it with "copy" in the name.

Add additional status codes that might be useful:
- F_bottom
- F_bottom_not
- F_character
- F_character_not
- F_fifo
- F_fifo_not
- F_middle
- F_middle_not
- F_regular
- F_regular_not
- F_top
- F_top_not

Provide a standard macro initializer and use it to make project maintenance easier.
This reduces lines of code in the programs.

Update the program flags:
- Use "in" and "out" instead of "to" and "from".
- Separate the "file" from the now "in" and "out".

Update the build settings files, making sure the locale and language environment variables are exposed by default.

Relocate the utf-8 code that previously didn't properly get properly relocated.