Refactor: Redesign fl_string and fl_utf_string, relocating into f_compare, f_parse, and f_rip.
The fl_string and fl_utf_string projects only depend on the core projects.
They could easily be moved into the level_0 project directory.
The problem is that they cannot be part of the f_string project due to depending on f_utf.
The f_string project is before the f_utf.

The solution is to break the fl_string into three projects:
- f_compare
- f_parse
- f_rip

It just so happens that the fl_utf_string is in the same situation as fl_string.
Move fl_utf_string into these three projects as appropriate.

This adds some initial unit tests to f_compare.
I didn't have enough to to finish f_compare, let alone get to f_parse and f_rip.