Bugfix: Remove extra arguments being added.
The parameter expansion is perhaps the worst designed area of the Featureless Make project.
It went through a lot of iterations and the logic requires jumping around too much.

At some point I changed the code to fix problems with the expansion logic regarding quotes and not quotes.
Two functions were left that still added arguments of their own.
This seems wrong.
This does not appear to cause an obvious bug and so it is also unclear if removing these cause a non-obvious regression.
Keep an eye out for regressions in this.