Bugfix: Fix bugs in console parameter processing exposed by unit tests.
The f_console_identify() function is incorrectly assigning the bits for short alone values.
The mistake is a missing pipe character and as a result the flags are being assigned rather than being bitwise-or appended.

The short.used and long.used string checks should skip when the string is empty.
This is not happening due to an AND comparitor in the string.
This then causes a false positive and the unit tests identify invalid data.

The process found check outside of the loop that depends on the iterator "i" should also check that i is less than the total array length.