Bugfix: Fix problems exposed by unit tests.
The f_pipe_debug and f_pipe_warning are no more and there is now an f_pipe_output.

The f_print_to* functions should not use parameter checking and should instead eturn F_file_descriptor_not when the descriptor is -1.

The F_help and F_help_not to string conversion is missing.