Progress: Make some progress towards getting fss_read up and running.
This mostly focuses on the print help functions.

Fix the incorrectly spelled "aplied".
Fix the help messages in the read man pages.

Add *_format_code_name_* for an upper-case human friendly name.
Add this to the fss_write as well given that the programs follow the "programs as a library" design.