Bugfix: Incorrect return code in tests causes false failures in CI.
Recent changes of the code change the "failure" variable from being an empty string for no error to being the explicit digit of "0".
The change is incomplete and this caused the false failure by incorrectly returning "1" when instead "0" should be returned.

The Gitlab CI detected this but the Github CI failed to.