Progress: Continue work on refactoring fss_*_read into fss_read.
This gets the programs compiling.
I decided not to get everything correct and instead decided to get everything compiling.

The memory allocation needs to be fixed.
There are a lot of fixme and todo comments to address.
The payload and embedded list need special handling and are very much not implemented.
Individual standard differences need to be handled (and some are stubbed out or copied over, making them invalid).